RWAHS Theme
There is a dependence on another themeVenuetheme/Yule, but if not
present will default back Magento/blank, so will mostly function. The
area that is having the most issues is Magento_Theme/layout/default.xml
and Magento_Theme/page_layout/1column.xml, this is where I’m trying to
reorganise the elements in the XML feed so that I can get a structure
similar to the website header.